requirements.txt
opencv-python==4.10.0.82
protobuf
decord==0.6.0
ipywidgets
ipykernel

wandb==0.19.11
ipykernel
transformers==4.57.1
nbformat==5.10.4
accelerate==1.11.0
datasets==4.1.1
bitsandbytes==0.48.2
peft==0.18.0
evaluate==0.4.4
# flash-attn==2.8.0.post2
scikit-learn==1.7.0
torchao==0.14.1
liger-kernel==0.6.2
torchtune==0.6.1
matplotlib==3.10.3
plotly==6.2.0
ipywidgets
more-itertools==10.8.0
seaborn==0.13.2
xformers==0.0.33
trl==0.25.1
lpips==0.1.4
clean-fid==0.1.35
vision-aided-loss==0.1.0
clip @ git+https://github.com/openai/CLIP.git
open-clip-torch==3.2.0
scikit-image==0.25.2
torchmetrics==1.8.2
piqa==1.3.2
jupyter-splitview==0.1.2
torch-fidelity==0.3.0
websocket-client==1.9.0
soundfile==0.13.1